+++ src/core/constants.cxx
@@ -1,4 +1,3 @@
-module;
 export module libfork.core:constants;
 
 import std;
