+++ src/core/core.cxx
@@ -7,10 +7,12 @@ export import :utility;
 export import :tuple;
 
 // T2 partitions
-export import :frame;   // concepts
+export import :frame;   // concepts, constants
 export import :context; // concepts
 export import :ops;     // concepts, tuple, utility
-export import :deque;   // concepts
+export import :deque;   // FIXME:
+
+export import :handles; // frame
 
 export import :schedule;
